move kzg methods to kzg.go
